// apps/server/src/generated/prisma/models/CrmEmailTemplateGroup.ts
export type CrmEmailTemplateGroupModel = runtime.Types.Result.DefaultSelection<Prisma.$CrmEmailTemplateGroupPayload>
// ...
export type $CrmEmailTemplateGroupPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "CrmEmailTemplateGroup"
  objects: {
    organization: Prisma.$OrganizationPayload<ExtArgs>
    steps: Prisma.$CrmEmailTemplateStepPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    organizationId: string
    name: string
    language: string
    description: string | null
    status: string
    isDefault: boolean
    createdById: string
    createdByName: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["crmEmailTemplateGroup"]>
  composites: {}
}