// apps/server/src/generated/prisma/models/CrmEmailTemplateStep.ts
export type CrmEmailTemplateStepModel = runtime.Types.Result.DefaultSelection<Prisma.$CrmEmailTemplateStepPayload>
// ...
export type $CrmEmailTemplateStepPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "CrmEmailTemplateStep"
  objects: {
    organization: Prisma.$OrganizationPayload<ExtArgs>
    templateGroup: Prisma.$CrmEmailTemplateGroupPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    organizationId: string
    templateGroupId: string
    stepIndex: number
    name: string
    threadMode: string
    delayDays: number
    subjectTemplate: string
    bodyTemplate: string
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["crmEmailTemplateStep"]>
  composites: {}
}